admin panel

+++ src/main/java/ru/shchetinin/lab1p/service/MovieService.java
@@ -7,16 +7,19 @@ import jakarta.transaction.Transactional;
 import jakarta.validation.Valid;
 import jakarta.ws.rs.core.Response;
 import ru.shchetinin.lab1p.dao.MovieDao;
+import ru.shchetinin.lab1p.dao.UserDao;
 import ru.shchetinin.lab1p.dto.request.MovieRequest;
 import ru.shchetinin.lab1p.entity.Coordinates;
 import ru.shchetinin.lab1p.entity.Movie;
 import ru.shchetinin.lab1p.entity.Person;
 import org.modelmapper.ModelMapper;
+import ru.shchetinin.lab1p.entity.Role;
 import ru.shchetinin.lab1p.excepion.EndpointException;
 
 import java.time.LocalDateTime;
 import java.util.Date;
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 @ApplicationScoped
@@ -33,12 +36,18 @@ public class MovieService {
     @Inject
     private CoordinatesService coordinatesService;
 
+    @Inject
+    private UserDao userDao;
 
     @Transactional
-    public Movie createMovie(MovieRequest movieRequest) {
+    public Movie createMovie(MovieRequest movieRequest, String username) {
         Movie movie = modelMapper.map(movieRequest, Movie.class);
         setDependencies(movieRequest, movie);
         movie.setCreationDate(new Date());
+
+        var opt = userDao.findByUsername(username);
+        opt.ifPresent(user -> movie.setCreator(user.getId()));
+
         movieDao.save(movie);
 
         return movie;
@@ -58,7 +67,20 @@ public class MovieService {
     }
 
     @Transactional
-    public Movie updateMovie(Long id, MovieRequest movieRequest) {
+    public Movie updateMovie(Long id, MovieRequest movieRequest, String username) {
+
+        var opt = userDao.findByUsername(username);
+        boolean isAuth = false;
+
+        if (opt.isPresent()) {
+            var creator = movieDao.findById(id).get().getCreator();
+            isAuth = Objects.equals(creator, opt.get().getId()) || opt.get().getRole().equals(Role.ADMIN);
+        }
+
+        if (!isAuth) {
+            throw new EndpointException(Response.Status.FORBIDDEN, String.format("Person with id %d is not creator", id));
+        }
+
         Optional<Movie> movieOptional = movieDao.findById(id);
         if (movieOptional.isEmpty()) {
             throw new EndpointException(Response.Status.NOT_FOUND, String.format("Movie with id %d not found", id));
@@ -73,7 +95,21 @@ public class MovieService {
     }
 
     @Transactional
-    public void deleteMovie(Long id) {
+    public void deleteMovie(Long id, String username) {
+
+        var opt = userDao.findByUsername(username);
+        boolean isAuth = false;
+
+        if (opt.isPresent()) {
+            var creator = movieDao.findById(id).get().getCreator();
+            isAuth = Objects.equals(creator, opt.get().getId()) || opt.get().getRole().equals(Role.ADMIN);
+        }
+
+        if (!isAuth) {
+            throw new EndpointException(Response.Status.FORBIDDEN, String.format("Person with id %d is not creator", id));
+        }
+
+
         if(!movieDao.delete(id)) {
             throw new EndpointException(Response.Status.NOT_FOUND, String.format("Movie with id %d not found", id));
         }
